+++ src/pgrls/fixers/__init__.py
@@ -66,6 +66,7 @@ def default_fixers() -> list[Fixer]:
     from pgrls.fixers.sec019 import SEC019Fixer
     from pgrls.fixers.sec020 import SEC020Fixer
     from pgrls.fixers.sec031 import SEC031Fixer
+    from pgrls.fixers.sec032 import SEC032Fixer
     from pgrls.fixers.view001 import VIEW001Fixer
     from pgrls.fixers.view002 import VIEW002Fixer
 
@@ -77,6 +78,7 @@ def default_fixers() -> list[Fixer]:
         SEC019Fixer(),
         SEC020Fixer(),
         SEC031Fixer(),
+        SEC032Fixer(),
         PERF001Fixer(),
         PERF003Fixer(),
         HYG003Fixer(),